♻️ refactor(lanyard): remove song progress tracking
going back to how it was before because it got too complicated with the logic, so i cba.

also switched mentions of 'emoji' to 'status dot' cos its a status dot, not an emoji anymore.

also added borders to few more things and made the song name the only clickable bit that takes you to the song url.