🐛 fix(anime): workaround for anime that don't have english titles
for anime that do not have an official English title or return `null` for the English title, add the AniList ID of the show to the `titleOverrides` object below, along with the preferred/correct title to display.

If there is no title override, it will use the English title (provided that it exists) by default, along with the Romaji and Native title just in case.